Moved navigation buttons to TripView
