+++ Pathways/app/src/main/java/com/example/pathways/NoteView.java
@@ -77,6 +77,7 @@ public class NoteView extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_note_view);
+        ((AppCompatActivity)this).getSupportActionBar().setDisplayHomeAsUpEnabled(false);
 
         ArrayList notes = new ArrayList<Note>();
