+++ Pathways/app/src/main/res/layout/activity_trip_view.xml
@@ -13,6 +13,7 @@
         android:layout_marginLeft="10dp"
         android:layout_marginRight="10dp"
         android:layout_marginTop="10dp"
+        android:layout_marginBottom="10dp"
         android:layout_height="wrap_content"
         android:clickable="true"
         android:focusable="true"
@@ -42,6 +43,36 @@
         app:layout_anchor="@id/locations_list" />
 
     <include layout="@layout/locations_list"/>
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/image_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/fab_margin_top"
+        android:layout_marginRight="@dimen/fab_margin_right"
+        app:backgroundTint="#161414"
+        android:layout_gravity="top|right|end"
+        app:srcCompat="@android:drawable/ic_menu_gallery" />
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/note_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/fab_margin_top"
+        android:layout_marginRight="@dimen/fab_margin_right"
+        app:backgroundTint="#161414"
+        app:srcCompat="@android:drawable/ic_menu_edit"
+        android:layout_gravity="top|right|end" />
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/menu_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/fab_margin_top"
+        android:layout_marginRight="@dimen/fab_margin_right"
+        app:backgroundTint="#161414"
+        app:srcCompat="@android:drawable/ic_input_add"
+        android:layout_gravity="top|right|end" />
 
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
