+++ Pathways/app/src/main/res/values/dimens.xml
@@ -1,5 +1,7 @@
 <resources>
-    <dimen name="fab_margin">16dp</dimen>
+    <dimen name="fab_margin">10dp</dimen>
+    <dimen name="fab_margin_top">80dp</dimen>
+    <dimen name="fab_margin_right">10dp</dimen>
     <dimen name="standard_60">60dp</dimen>
     <dimen name="standard_120">120dp</dimen>
     <dimen name="standard_180">180dp</dimen>
